// mlflow/server/js/src/shared/web-shared/genai-traces-table/types.ts
export type AssessmentDType = 'string' | 'numeric' | 'boolean' | 'pass-fail' | 'unknown';
export type AssessmentType = 'AI_JUDGE' | 'HUMAN' | 'CODE';

// Reflects structure logged by mlflow.log_table()
export interface RawGenaiEvaluationArtifactResponse {
  columns?: (string | null)[];
  data?: (string | number | null | boolean | Record<string, any>)[][];
  filename: string;
}

export interface AssessmentInfo {
  name: string;
  displayName: string;
  // True when the assessment comes from a built-in judge.
  isKnown: boolean;
  isOverall: boolean;
  // The metric that produced the assessment. Defined as the name of the judge or custom metric function that produced the assessment.
  metricName: string;
  source?: RunEvaluationResultAssessmentSource;
  isCustomMetric: boolean;
  isEditable: boolean;
  isRetrievalAssessment: boolean;
  // The type of the assessment value.
  dtype: AssessmentDType;
  uniqueValues: Set<AssessmentValueType>;

  // Display metadata.
  docsLink: string;
  missingTooltip: string;
  description: string;

  // True if if the assesment contains at least one error
  containsErrors?: boolean;
}
// ...
export type RunEvaluationResultAssessmentSource = {
  sourceType: AssessmentType;
  sourceId: string;
  metadata: Record<string, string>;
};
// ...
export type AssessmentValueType = string | boolean | number | undefined;
// ...
export interface EvaluationsOverviewTableSort {
  key: string;
  type: TracesTableColumnType;
  asc: boolean;
}

  export interface TraceActions {
    exportToEvals?: {
      showExportTracesToDatasetsModal: boolean;
      setShowExportTracesToDatasetsModal: (visible: boolean) => void;
      renderExportTracesToDatasetsModal: ({
        selectedTraceInfos,
        onSuccess,
      }: {
        selectedTraceInfos: ModelTraceInfoV3[];
        onSuccess?: (args: {
          experimentId: string;
          datasetId: string;
          datasetName: string;
          count: number;
        }) => void;
      }) => React.ReactNode;
    };
  deleteTracesAction?: {
    deleteTraces?: (experimentId: string, traceIds: string[]) => Promise<any>;
    isDisabled?: boolean;
    disabledReason?: string;
  };
  editTags?: {
    showEditTagsModalForTrace: (trace: ModelTraceInfoV3) => void;
    EditTagsModal: React.ReactNode;
  };
}

// @deprecated, use TableFilter instead
export interface AssessmentFilter {
  assessmentName: string;
  filterValue: AssessmentValueType;
  // Only defined when filtering on an assessment for RCA values.
  filterType?: 'rca' | undefined;
  run: string;
}
export type TableFilter = {
  // The column group (e.g. "Assessments") or a specific column (e.g. "execution_duration")
  column: TracesTableColumnGroup | string;
  // Should be defined if a column group is used.
  key?: string;
  operator: FilterOperator | HiddenFilterOperator;
  value: TableFilterValue;
};

export type TableFilterValue = string | boolean | number | undefined;

export interface TableFilterOption {
  value: string;
  renderValue: () => string | React.ReactNode;
}

export interface TableFilterOptions {
  source: TableFilterOption[];
}

export enum FilterOperator {
  EQUALS = '=',
  GREATER_THAN = '>',
  LESS_THAN = '<',
  GREATER_THAN_OR_EQUALS = '>=',
  LESS_THAN_OR_EQUALS = '<=',
}

// operators that are not displayed in the filter popover, but are
// still supported in the backend. eventually we should implement
// functionality for all these operators, but some of them take a
// little more thought from the UX side (e.g. we need to remove the
// value input for IS NOT NULL filters)
export enum HiddenFilterOperator {
  ILIKE = 'ILIKE',
}
// ...
export enum TracesTableColumnType {
  ASSESSMENT = 'ASSESSMENT',
  EXPECTATION = 'EXPECTATION',
  TRACE_INFO = 'TRACE_INFO',
  INPUT = 'INPUT',
  // This is a hack so that internal agent monitoring can display request time.
  INTERNAL_MONITOR_REQUEST_TIME = 'INTERNAL_MONITOR_REQUEST_TIME',
}

// This represents columns that are grouped together.
// For example, each assessment is its own column, but they are all grouped under the "Assessments" column group.
export enum TracesTableColumnGroup {
  ASSESSMENT = 'ASSESSMENT',
  EXPECTATION = 'EXPECTATION',
  TAG = 'TAG',
  INFO = 'INFO',
}

export const TracesTableColumnGroupToLabelMap = {
  [TracesTableColumnGroup.ASSESSMENT]: 'Assessments',
  [TracesTableColumnGroup.EXPECTATION]: 'Expectations',
  [TracesTableColumnGroup.TAG]: 'Tags',
  // We don't show a label for the info column group
  [TracesTableColumnGroup.INFO]: '\u00A0',
};

export interface TracesTableColumn {
  /** This is the assessment name for assessments, and a static string for trace info and input columns */
  id: string;
  /** The label for the column, displayed in the table header. */
  label: string;
  /** The label for the column used in filter dropdowns. If not provided, defaults to `label`. */
  filterLabel?: string;
  type: TracesTableColumnType;
  group?: TracesTableColumnGroup;

  // TODO: Remove this field once migration to trace info v3 is complete
  assessmentInfo?: AssessmentInfo;
  expectationName?: string;
}